Move Packet capturing stuff to pktcapture package.
